Update decommission-enterprise-certification-authority-and-remove-objects.md
I have just used this article to decomm an Enterprise CA and found a number of syntax and clarity issues with it.
I have updated it with my corrections and additions.
Note that the formatting seems to be a bit screwy (at least from what I can see in the preview) from Step 6 onwards, with some sub-numbering, console, note and important sections not displaying correctly.  I tried to fix it but was unable to do so.